+++ ChangeLog.md
@@ -9,7 +9,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * `Starsystem` object updated
       * `ContestingPowers` and `contestingpowers` updated. These are now lists of powerplay powers with at least 30% progress towards acquisition of the star system, if any, either as objects or as strings, in descending order.
       * `NearbyPowers` and `nearbypowers` added. These are lists of powerplay powers having star systems with acquisition radii which overlap the star system, either as objects or as strings.
-      * `powerAcquisitionProgress` added. This is the progress of nearby powerplay powers towards obtaining control of the star system, as a list of objects where the key is the power (as an object) and the value is the percent progress towards achieving the control score threshold.
+      * `powerAcquisitionProgress` added. This is the progress of nearby powerplay powers towards obtaining control of the star system, as a list of objects with keys Power (as an object) and progress (as a percent).
       * `powerControlProgress` added. This is the percent progress of the controlling power, if any, in consolidating control over the star system. Values below 0% indicate a reduction in the control level at the end of the cycle while values above 100% indicate an increase in the control level at the end of the cycle (if the current control state is less than 'Stronghold').
       * `powerReinforcementControlPoints` added. This is control points accumulated by the controlling power via powerplay reinforcement activities during the current cycle.
       * `powerUnderminingControlPoints` added. This is the control points lost by the controlling power via powerplay undermining activities during the current cycle.